+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * [JANUS-257] Fixed the type of subarray_id in the StationMetadata class to match with StationData.h
 * [THORN-257] Added MarkDone() command
 * [JANUS-273] Updated environment variable used in acquire_station_beam.cpp to align with other data modes
+* [THORN-259] Implementation of alarms on Ringbuffer occupancy, which maps to healthstate via ADR-115.
 
 ## 4.1.1